Baseline
This is the startine point for the modifications I'll make to the FAST
control system subroutines. The /source folder contains the unmodified
FAST V7.02 source code.  /Compiling contains scripts for compiling FAST
with Intel Visual Fortran as well as a few precompiled versinos of FAST.
/DLLs contains the source code for several Bladded style controllers and
a script for compiling the source code into Dynamic Link Libraries
(DLL).