Make derate parameters primary.fst inputs
Modify code so derate start time, end time, and derate factor are
specified in the input file instead of in the UserSubs file. This way
FAST doesn't have to be recompiled to change the derating. A sample
input file primary.fst with the derate input parameters has been added
to the Compiling folder.